+++ packages/frontend/src/components/AboutFooter.tsx
@@ -17,7 +17,8 @@ import { IconBook, IconInfoCircle } from '@tabler/icons-react';
 import { FC, useState } from 'react';
 import { FOOTER_HEIGHT, FOOTER_MARGIN } from '../constants';
 import { useApp } from '../providers/AppProvider/useApp';
-import { TrackPage, TrackSection } from '../providers/TrackingProvider';
+import { TrackPage } from '../providers/TrackingProvider/TrackPage';
+import { TrackSection } from '../providers/TrackingProvider/TrackSection';
 import Logo from '../svgs/grey-icon-logo.svg?react';
 import { PageName, PageType, SectionName } from '../types/Events';
 import MantineIcon from './common/MantineIcon';
