+++ packages/frontend/src/components/Explorer/ResultsCard/CellContextMenu.tsx
@@ -18,7 +18,7 @@ import { useParams } from 'react-router-dom';
 import useToaster from '../../../hooks/toaster/useToaster';
 import { useFilters } from '../../../hooks/useFilters';
 import { useApp } from '../../../providers/AppProvider/useApp';
-import { useTracking } from '../../../providers/TrackingProvider';
+import { useTracking } from '../../../providers/TrackingProvider/useTracking';
 import { EventName } from '../../../types/Events';
 import { Can } from '../../common/Authorization';
 import MantineIcon from '../../common/MantineIcon';
