+++ packages/frontend/src/components/NavBar/GlobalSearch/index.tsx
@@ -26,7 +26,7 @@ import { useHistory, useLocation } from 'react-router-dom';
 import { GLOBAL_SEARCH_MIN_QUERY_LENGTH } from '../../../hooks/globalSearch/useGlobalSearch';
 import { useProject } from '../../../hooks/useProject';
 import { useValidationUserAbility } from '../../../hooks/validation/useValidation';
-import { useTracking } from '../../../providers/TrackingProvider';
+import { useTracking } from '../../../providers/TrackingProvider/useTracking';
 import { EventName } from '../../../types/Events';
 import MantineIcon from '../../common/MantineIcon';
 import { SearchItem, useDebouncedSearch } from './hooks';
