+++ packages/frontend/src/components/ProjectAccess/ProjectAccessCreation/index.tsx
@@ -8,11 +8,12 @@ import {
 import { Button, Group, Modal, ModalProps, Select, Title } from '@mantine/core';
 import { useForm } from '@mantine/form';
 import { IconUser } from '@tabler/icons-react';
-import React, { FC, useEffect, useState } from 'react';
+import { FC, useEffect, useState } from 'react';
 import { useCreateInviteLinkMutation } from '../../../hooks/useInviteLink';
 import { useOrganizationUsers } from '../../../hooks/useOrganizationUsers';
 import { useCreateProjectAccessMutation } from '../../../hooks/useProjectAccess';
-import { TrackPage, useTracking } from '../../../providers/TrackingProvider';
+import { TrackPage } from '../../../providers/TrackingProvider/TrackPage';
+import { useTracking } from '../../../providers/TrackingProvider/useTracking';
 import {
     CategoryName,
     EventName,
