+++ packages/frontend/src/components/ProjectConnection/ProjectConnectFlow/ConnectManually/ConnectManuallyStep1.tsx
@@ -2,7 +2,7 @@ import { Button, Stack, Text, Tooltip } from '@mantine/core';
 import { Prism } from '@mantine/prism';
 import { IconChevronLeft, IconChevronRight } from '@tabler/icons-react';
 import { FC } from 'react';
-import { useTracking } from '../../../../providers/TrackingProvider';
+import { useTracking } from '../../../../providers/TrackingProvider/useTracking';
 import { EventName } from '../../../../types/Events';
 import MantineIcon from '../../../common/MantineIcon';
 import { ProjectCreationCard } from '../../../common/Settings/SettingsCard';
