+++ packages/frontend/src/components/ProjectConnection/index.tsx
@@ -33,7 +33,7 @@ import {
 } from '../../hooks/useProject';
 import { useActiveJob } from '../../providers/ActiveJobProvider/useActiveJob';
 import { useApp } from '../../providers/AppProvider/useApp';
-import { useTracking } from '../../providers/TrackingProvider';
+import { useTracking } from '../../providers/TrackingProvider/useTracking';
 import { EventName } from '../../types/Events';
 import { useAbilityContext } from '../common/Authorization';
 import MantineIcon from '../common/MantineIcon';
