+++ packages/frontend/src/components/SimpleTable/DashboardCellContextMenu.tsx
@@ -20,7 +20,7 @@ import { useParams } from 'react-router-dom';
 import useToaster from '../../hooks/toaster/useToaster';
 import { useApp } from '../../providers/AppProvider/useApp';
 import { useDashboardContext } from '../../providers/DashboardProvider/useDashboardContext';
-import { useTracking } from '../../providers/TrackingProvider';
+import { useTracking } from '../../providers/TrackingProvider/useTracking';
 import { EventName } from '../../types/Events';
 import { Can } from '../common/Authorization';
 import MantineIcon from '../common/MantineIcon';
