+++ packages/frontend/src/components/common/MantineLinkButton.tsx
@@ -1,7 +1,8 @@
 import { Anchor, Button, ButtonProps } from '@mantine/core';
 import React, { FC } from 'react';
 import { useHistory } from 'react-router-dom';
-import { EventData, useTracking } from '../../providers/TrackingProvider';
+import { EventData } from '../../providers/TrackingProvider/types';
+import { useTracking } from '../../providers/TrackingProvider/useTracking';
 
 export interface MantineLinkButtonProps extends ButtonProps {
     href: string;
