+++ packages/frontend/src/features/tableCalculation/components/UpdateTableCalculationModal.tsx
@@ -2,7 +2,7 @@ import { TableCalculation } from '@lightdash/common';
 import { ModalProps } from '@mantine/core';
 import { FC } from 'react';
 import { useExplorerContext } from '../../../providers/ExplorerProvider/useExplorerContext';
-import { useTracking } from '../../../providers/TrackingProvider';
+import { useTracking } from '../../../providers/TrackingProvider/useTracking';
 import { EventName } from '../../../types/Events';
 import TableCalculationModal from './TableCalculationModal';
