+++ packages/frontend/src/pages/Login.tsx
@@ -30,7 +30,7 @@ import { ThirdPartySignInButton } from '../components/common/ThirdPartySignInBut
 import PageSpinner from '../components/PageSpinner';
 import useToaster from '../hooks/toaster/useToaster';
 import { useApp } from '../providers/AppProvider/useApp';
-import { useTracking } from '../providers/TrackingProvider';
+import { useTracking } from '../providers/TrackingProvider/useTracking';
 import LightdashLogo from '../svgs/lightdash-black.svg';
 
 type LoginParams = { email: string; password: string };
